+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added `PhysicalAddress` and `VirtualAddress` type aliases.
 - Added `Guid::from_bytes` and `Guid::to_bytes`.
+- Added `UnalignedSlice` for representing a reference to an unaligned
+  slice.
+- Added `DeviceSubType::MESSAGING_REST_SERVICE` and
+  `DeviceSubType::MESSAGING_NVME_OF_NAMESPACE`.
 
 ### Changed
 
@@ -22,6 +26,12 @@
 - The `Revision` type now implements `Display` with correct formatting
   for all UEFI versions. The custom `Debug` impl has been removed and
   replaced with a derived `Debug` impl.
+  
+### Removed
+
+- Removed `UnalignedCStr16`; use `UnalignedSlice` instead. An
+  `UnalignedSlice<u16>` can be converted to a string with `to_cstr16` or
+  `to_cstring16`.
 
 ## uefi-macros - [Unreleased]